Refactor amphitheatre into modules
