+++ CHANGELOG_ARCHIVE.md
@@ -1,5 +1,11 @@
 # Changelog Archive
 
+## 2025-07-23
+- 1621 Added "Use Animated Sprites" option with new sprite models and animations
+- 2214 Break Game.js into smaller initialization modules for easier maintenance
+- 1614 Move ground textures into assets/ground_textures and update references
+- 0000 Rewrite mats.json to load building modules from `js/mats`
+
 ## 2025-07-22
 - 2123 Add distinct biomes and tree types. The world now features five biomes: a central Grassland, Sand (Palm trees), Snow (Snowy Pine trees), Forest (Deciduous and Pine trees), and Dirt/Stone (Sparse, bare trees).
 - 2122 Add music button and modal with scrollable song and collaborator lists
